refactor: :technologist: refactor all views
improve legibility, fix typos
