+++ app/templates/app/category/categories.html
@@ -1,0 +1,70 @@
+{% extends "base.html" %}
+
+{% block title %}Categoría{% endblock %}
+
+{% block content %}
+<div class="container">
+    <div class="d-flex justify-content-between align-items-center mb-4">
+        <h1>Categoría</h1>
+            <a
+                href="{% url 'category_form' %}"
+                class="btn btn-primary"
+            >
+                <i class="bi bi-plus-circle me-2" aria-hidden="true"></i>
+                Crear Categoría
+            </a>
+    </div>
+    <table class="table">
+        <thead>
+            <tr>
+                <th>Título</th>
+                <th>Dirección</th>
+                <th>Ciudad</th>
+                <th>Capacidad</th>
+                <th>Contacto</th>
+            </tr>
+        </thead>
+        <tbody>
+            {% for category in categorys%}
+                <tr>
+                    <td>{{ category.name }}</td>
+                    <td>{{ category.address }}</td>
+                    <td>{{ category.city }}</td>
+                    <td>{{ category.capacity }}</td>
+                    <td>{{ category.contact }}</td>
+                    <td>
+                        <div class="hstack gap-1">
+                            <a href="{% url 'category_detail' category.id %}"
+                               class="btn btn-sm btn-outline-primary"
+                               aria-label="Ver detalle"
+                               title="Ver detalle">
+                                <i class="bi bi-eye" aria-hidden="true"></i>
+                            </a>
+                                <a href="{% url 'category_edit' category.id %}"
+                                    class="btn btn-sm btn-outline-secondary"
+                                    aria-label="Editar"
+                                    title="Editar">
+                                    <i class="bi bi-pencil" aria-hidden="true"></i>
+                                </a>
+                                <form action="{% url 'category_delete' category.id %}" method="POST">
+                                    {% csrf_token %}
+                                    <button class="btn btn-sm btn-outline-danger"
+                                        title="Eliminar"
+                                        type="submit"
+                                        aria-label="Eliminar"
+                                        titile="Eliminar">
+                                        <i class="bi bi-trash" aria-hidden="true"></i>
+                                    </button>
+                                </form>
+                        </div>
+                    </td>
+                </tr>
+            {% empty %}
+                <tr>
+                    <td colspan="4" class="text-center">No hay Categoríaes creados</td>
+                </tr>
+            {% endfor %}
+        </tbody>
+    </table>
+</div>
+{% endblock %}
