+++ app/urls.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.views import LogoutView
 from django.urls import path
 
-from . import views
+from app import views
 
 urlpatterns = [
     path("", views.home, name="home"),
@@ -20,10 +20,10 @@ urlpatterns = [
     path("events/<int:id>/edit/", views.event_form, name="event_edit"),
     path("events/<int:id>/", views.event_detail, name="event_detail"),
     path("events/<int:id>/delete/", views.event_delete, name="event_delete"),
-    path("refunds/", views.My_refunds, name="my_refunds"),
-    path("refunds/new/", views.Refund_create, name="refund_create"),
-    path("refunds/<int:id>/edit/", views.Refund_edit, name="refund_edit"),
-    path("refunds/<int:id>/delete/", views.Refund_delete, name="refund_delete"),
+    path("refunds/", views.my_refunds, name="my_refunds"),
+    path("refunds/new/", views.refund_create, name="refund_create"),
+    path("refunds/<int:id>/edit/", views.refund_edit, name="refund_edit"),
+    path("refunds/<int:id>/delete/", views.refund_delete, name="refund_delete"),
     path("organizer/refund/", views.RefundRequestsAdminView.as_view(), name="refunds_admin"),
     path("organizer/refund/aprobar/<int:pk>/", views.approve_refund_request, name="refund_approve"),
     path("organizer/refund/rechazar/<int:pk>/", views.reject_refund_request, name="refund_reject"),
