+++ app/views/__init__.py
@@ -1,0 +1,6 @@
+from .auth_views import *
+from .event_views import *
+from .venue_views import *
+from .category_views import *
+from .refund_views import *
+from .rating_views import *
