+++ app/views/auth_views.py
@@ -1,0 +1,54 @@
+from django.contrib.auth import authenticate, login
+from django.shortcuts import redirect, render
+
+from app.models import User
+
+def register(request):
+    if request.method == "POST":
+        email = request.POST.get("email")
+        username = request.POST.get("username")
+        is_organizer = request.POST.get("is-organizer") is not None
+        password = request.POST.get("password")
+        password_confirm = request.POST.get("password-confirm")
+
+        errors = User.validate_new_user(email, username, password, password_confirm)
+
+        if len(errors) > 0:
+            return render(
+                request,
+                "accounts/register.html",
+                {
+                    "errors": errors,
+                    "data": request.POST,
+                },
+            )
+        else:
+            user = User.objects.create_user(
+                email=email, username=username, password=password, is_organizer=is_organizer
+            )
+            login(request, user)
+            return redirect("events")
+
+    return render(request, "accounts/register.html", {})
+
+
+def login_view(request):
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is None:
+            return render(
+                request, "accounts/login.html", {"error": "Usuario o contraseña incorrectos"}
+            )
+
+        login(request, user)
+        return redirect("events")
+
+    return render(request, "accounts/login.html")
+
+
+def home(request):
+    return render(request, "home.html")
